添加删除数据接口

+++ src/main/java/zyxhj/jiti/domain/ORGUser.java
@@ -57,7 +57,7 @@ public class ORGUser {
 	/**
 	 * 投票权重
 	 */
-	@RDSAnnField(column = RDSAnnField.INTEGER)
+	@RDSAnnField(column = RDSAnnField.DOUBLE)
 	public Double weight;
 	
 	/**
